requirements.txt
agentops
python-dotenv
openai
git+https://github.com/AgentOps-AI/crewAI.git@main